HospitalParser 작성 중

+++ src/main/java/com/practice/hospitalapi/domain/Hospital.java
@@ -27,5 +27,5 @@ public class Hospital {
     private int healthcareProviderCount ;
     private int patientRoomCount;
     private int totalNumberOfBeds;
-    private double totalAreaSize;
+    private float totalAreaSize;
 }
